Update readme.txt tags to include 'woo'
As discussed in p6q8Tx-1qz-p2, when users search for 'woo' on the WP.org plugin repository, WooCommerce does not appear in the list of results. This is happening because 'woo' is not mentioned in the content, excerpt, title, or tags. It was suggested that we start using the tags fields for synonyms like 'woo'. Since WP.org only stores the first 5 tags, in this commit, I'm removing the tag ecommerce tag (which is a word that appears both in the excerpt and the content) and replacing it with 'woo'.

In the future, we might want to evaluate the other four tags that WP.org is considering: e-commerce, store, sales, sell.